backend/requirements.txt
fastapi
uvicorn
python-multipart
replicate
python-dotenv
